Remove Extensions namespace #14
